Upgrade to ETP1.2 Dev15
Upgrade to gsoap 2.8.106
Code cleaning (less warning)

+++ src/resqml2/EarthModelInterpretation.cpp
@@ -60,17 +60,17 @@ void EarthModelInterpretation::loadTargetRelationships()
 	}
 	
 	unsigned int count = getStructuralOrganizationInterpretationCount();
-	for (size_t i = 0; i < count; ++i) {
+	for (unsigned int i = 0; i < count; ++i) {
 		convertDorIntoRel<StructuralOrganizationInterpretation>(getStructuralOrganizationInterpertationDor(i));
 	}
 
 	count = getStratiOccurrenceCount();
-	for (size_t i = 0; i < count; ++i) {
+	for (unsigned int i = 0; i < count; ++i) {
 		convertDorIntoRel<StratigraphicOccurrenceInterpretation>(getStratiOccurrenceDor(i));
 	}
 
 	count = getRockFluidOrganizationInterpretationCount();
-	for (size_t i = 0; i < count; ++i) {
+	for (unsigned int i = 0; i < count; ++i) {
 		convertDorIntoRel<RockFluidOrganizationInterpretation>(getRockFluidOrganizationInterpretationDor(i));
 	}
 }
